Ensure the tzdata package is installed on Ubuntu Jammy
This package would be installed on an actual Ubuntu Jammy system per
the manifests:
https://releases.ubuntu.com/jammy/ubuntu-22.04.4-desktop-amd64.manifest
https://releases.ubuntu.com/jammy/ubuntu-22.04.4-live-server-amd64.manifest

However, it is absent on our Docker image and thus we are missing
timezone data. This causes the geoipupdate timer we add to fail when
it is loaded. I verified this was the problem by logging into the
container and when I ran `timedatectl list-timezones` the only result
was "UTC". Once I installed the tzdata package the same command
resulted in an expected list of timezones.